+++ app/components/base/alert/component.html.erb
