+++ app/views/users/passwords/edit.html.erb
@@ -1,4 +1,6 @@
-<h2>Change your password</h2>
+<%= render Header::Component.new(image: image_url("header-image.jpeg"),
+                                 heading: "Update profile",
+                                 subheading: "Change Password") %>
 
 <%= form_for(resource, as: resource_name, url: password_path(resource_name), html: { method: :put }) do |f| %>
   <%= render "users/shared/error_messages", resource: resource %>
