+++ app/views/users/passwords/new.html.erb
@@ -1,4 +1,6 @@
-<h2>Forgot your password?</h2>
+<%= render Header::Component.new(image: image_url("header-image.jpeg"),
+                                 heading: "Forgot your password?",
+                                 subheading: "Reset password") %>
 
 <%= form_for(resource, as: resource_name, url: password_path(resource_name), html: { method: :post }) do |f| %>
   <%= render "users/shared/error_messages", resource: resource %>
