+++ app/views/users/registrations/edit.html.erb
@@ -1,3 +1,7 @@
+<%= render Header::Component.new(image: image_url("header-image.jpeg"),
+                                 heading: "Did some mistakes",
+                                 subheading: "Edit profile") %>
+
 <div class="d-flex justify-content-center">
   <div class="col-4">
 <h2>Edit <%= resource_name.to_s.humanize %></h2>
