+++ app/views/users/registrations/new.html.erb
@@ -1,3 +1,7 @@
+<%= render Header::Component.new(image: image_url("header-image.jpeg"),
+                                 heading: "Want to write a blog?",
+                                 subheading: "Register now") %>
+
 <div class="d-flex justify-content-center">
   <div class="col-4">
     <h2>Sign up</h2>
