+++ app/views/users/sessions/new.html.erb
@@ -1,6 +1,8 @@
+<%= render Header::Component.new(image: image_url("header-image.jpeg"),
+                                 heading: "Login",
+                                 subheading: "Write a blog") %>
 <div class="d-flex justify-content-center">
   <div class="col-4">
-    <h2>Log in</h2>
     <%= form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>
       <div class="field">
         <%= f.label :email %><br />
